+++ CHANGELOG.md
@@ -27,6 +27,8 @@
 
 ### Changed
 - Updated minimal theme
+- Update translations (03.06.2022)
+- New login screen layout
 - Update translations (07.06.2022)
 - Update deltachat-node to `1.86.0`